[Refactor] renamed Ratings component to Stars
