+++ client/public/src/components/RatingsBox.jsx
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+
+const RatingsBox = () => {
+  return (
+    <div className={StyleSheet.ratingsContainer}>
+      <div className={styles.ratingsContainerInner}>
+        <div>
+        </div>
+      </div>
+    </div>
+  );
+};
