+++ client/public/src/styles/ratings.styles.css
@@ -1,0 +1,18 @@
+.ratingsContainer {
+  margin-left: -8px !important;
+  margin-right: -8px !important;
+}
+
+.ratingsContainer::before {
+  content: " " !important;
+  display: table !important;
+}
+
+.ratingerContainerInner {
+  width: 50% !important;
+  float: left !important;
+  padding-left: 8px !important;
+  padding-right: 8px !important;
+  min-height: 1px !important;
+  position: relative !important;
+}
